TIP-1287: move ComputeEntityRawValuesSubscriber
